XMLHttpRequest/tests/submissions/Microsoft/abort-event-loadend.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The abort() method: Fire a progress event named loadend</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function()
        {
            var xhr = new XMLHttpRequest();
            var initialized = false;

            xhr.onreadystatechange = function()
            {
                test.step(function()
                {
                    if (xhr.readyState == 1) 
                    {
                        if (initialized) 
                        {
                            xhr.abort();
                        }
                        initialized = true;
                    }
                });
            };

            xhr.onloadend = function(e)
            {
                test.step(function() 
                {
                    assert_true(e instanceof ProgressEvent);
                    assert_equals(e.type, "loadend");
                    test.done(); 
                });
            };

            xhr.open("GET", "./support/echo.php", true);
            xhr.send();
        });
    </script>
</body>
</html>
